curriculum_sequencing_rl/requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0

# Visualization and analysis
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=5.4.0
dataclasses-json>=0.5.0

# Optional: For enhanced performance
# numba>=0.56.0  # Uncomment for JIT compilation
# scikit-learn>=1.0.0  # Uncomment for additional ML utilities
